sql/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
alembic
pytest
python-dotenv